[#108] Add version attribute to icat columns
In the models, new icat columns can now be declared with an optional min_version parameter to indicate the version of iRODS in which they first appear.
When building a query from a model, e.g: query(Resource), the columns used are those whose min_version attribute is lesser or equal to the current session's server version.