[#220] Replace stale connections pulled from idle pools
When getting a connection, check when the connection was last used
(This is done via connection's last_used_time field). If it was
last used more than a specific amount (configurable), drop the
connection and create a one. This is to address the issue with
connections that are dropped after being not used for a long time.