[#269] cleanup() is now automatic with session destruct.
Also: In cleanup( ), the session pool's connections will now be
properly shut down and the pool itself re-initialized to contain no
active or idle connections.  This is important in case the session
is used again, as happens in certain tests (see the admin_test) which
deliberately do this kind of refresh to update information from the
catalog.

When called from iRODSSession's __del__ destructor, an exception to
that rule is made, and the pool reinitialization does not happen.