[#290] only anonymous user can log in without password
Logins without a password (ie. when password='', or when the secrets file
".irodsA" is missing from the irods environment) should only be allowed for the
anonymous user.

The need for this fix was discovered when the demonstration script at:

  https://github.com/irods/python-irodsclient/issues/315#issue-1059980335

ceased to trip the issue at hand (#315) when rebased to follow commit:

  91b1adaa7c2e78c867407cd55008fe4c75de0ed5
  ( [#290] allow skipping of password file in anonymous user case )

Note that prior to the two #290 related commits (ie. this and the above-
mentioned one), instantiating an iRODSSession object from the user's
client environment without the secrets file present would raise a
FileNotFoundError.

As of this change, attempts by any other user than `anonymous' to log in
without auth credentials will instead raise an Exception of type:

  `NonAnonymousLoginWithoutPassword'