[#334] add SECURE_XML to parser selection
QUASI_XML has recently been added to the library along with
a mechanism for selecting the active parser for the current
thread.  We now make the "defusedxml" module into an option,
and it is selectable via:

  from irods.message import *
  ET( XML_Parser_Type.SECURE_XML )

We may now call

  ET( None )

at any time to reset any such changes made to the current thread.

Additionally, there are now two environment variables which may
be used to affect the XML parser used as the global default:

   - PYTHON_IRODSCLIENT_DEFAULT_XML: set to 'STANDARD_XML',
     'QUASI_XML',or 'SECURE_XML'.
   - PYTHON_IRODSCLIENT_QUASI_XML_SERVER_VERSION: set to '4,2,8'
     to adapt the 'QUASI_XML' parser for use with servers 4.2.8
     and prior.  The default is to use entities consistent with
     fixes made since 4.2.9 release.