[#372] eliminate SyntaxWarning ("is" operator being used with a literal)
In the no-argument form of the call to function irods.message.ET (which the
client executes quite often), the comparison against a literal 0 was only
being used to test whether an call parameter was allowed to default; but the
first-ever such call (just after the production of the .pyc containing the
Python3 byte-code) was then emitting the disconcerting SyntaxWarning. Here, we
change the parameter's default value to an empty tuple and then use a ==
comparison instead to suppress the SyntaxWarning and, with that, better conform
to Python3's expectations in terms of restricted uses of the 'is' operator.