[#369] remove dynamic generation of message classes
Revert to a single version of each of the classes:
    - MetadataRequest
    - TicketAdminRequest class
since altering the definition of these PackStruct message classes based on iRODS server
version turns out to be unnecessary.  The KeyValPair element of the message, used to
convey the ADMIN_KW value for newer iRODS servers, is actually ignored by the older
servers.

Also, skip the metadata and ticket tests of the newer ADMIN_KW functionality
if the detected server version pre-dates that keyword.