[#396] Introduce "acls" manager and deprecate "permissions"
The deprecated access manager, <session>.permissions, will be removed in v2.0.0.

Currently, either of <session>.permissions.get() or <session>.acls.get()
may be used and are identical, except that the latter lacks the
report_raw_acls parameter.

This parameter, when False, showed internal inconsistencies.  For
collections (but not for data objects) it reported more than was recorded
in the object catalog (#396). That made for unbalanced / surprising behaviors,
on a couple of fronts, such as are never welcome for application programmers.