[#418] raise error in test for IRODS_VERSION mismatch
We now probe the server_version in each test and quit if the
server is too recent, a strong reminder to update the IRODS_VERSION
in the StartupPack.

That probing initially caused test_failed_connection not to pass, due
to its implicit assumption of an empty idle pool - also fixed in this commit.