[#434] metadata calls now require AVU fields to be nonzero-length strings
AVU add and set operations will henceforth require attribute and value arguments
to be nonzero-length and exactly of type str (ie. Python string).  Only the units
field may have a value that is zero-length or, equivalently, of type None.
(Semantically, either one is the same as omitting the units.)

If an irods.meta.iRODSMeta object is used in the call, the above restrictions
apply to the corresponding fields of that object.