[#455] introduce low level ticket api changes
This ensures the application of a ticket to a session via supply()
filters down to each connection instantiated for that session object.

One test (test_object_read_and_write_tickets) was updated to ensure
call to cleanup() on a session, in order to clear out existing connections
(with their associated tickets) when required.  Due to the flawed approach
previously used in the ticket API, this step had not been strictly
necessary for the test to pass.