[#462][#399] Default the dataSize to 0 and prefer DATA_SIZE_KW over seek
Seeks are otherwise used to calculate object size.

The dataSize default of 0 is needed in servers before 4.3.1  because
otherwise -1 can wrap to (1L<<64)-1.

Also, we now send a positive dataSize parameter in a "put" to hint the
eventual size of a data objects so that a resource's minimum free space requirement is heeded.