[state of supporting PR #772 as of the current Pull Request's creation]
[_697] Github workflows for running PRC tests

Co-authored-by: Terrell Russell <tgr@renci.org>

Includes a docker compose configuration in which the PRC test suite
can be run in a way representative of typical operational use, and the
version of iRODS server and python interpreter that we install are also easily
reconfigurable.

Additionally, this affords us the opportunity to run the test suite on a client
node all its own, the iRODS server being reachable only via the network.

[_502] test harness and container-based tests

A new test harness is introduced in which we construct a new
container (using either Docker or podman) for each test program
we run. This allows full customization of the container
environment for the particular needs of each test.

Accordingly, included with the Github workflows (in a separate
commit) is a full run of the PRC test suite with the iRODS server
and catalog DB server running in the same container as the client.

In the process of putting old tests through new rigors, faults
were found and corrected in some of those tests.

completely remove {send,recv}_oneshot sync mechanism

review comments

Comments moved to *.yml files.  README.md no longer needed.

update irods/harness/README

deindent

correct forgotten comment symbol

clearer comment in run_tests.sh

suggestions nr. 1 10-22-25 13:19

suggestions nr. 2 10-22-25

smchg

package repo chg

comments chg