[#566] Rename login_auth_test.py to prevent running with full suite
The runner.py script which is used to run the full test suite looks for
files with the pattern "*_test.py". login_auth_test.py requires a special
environment and so should not be run with the regular suite of tests. It
must be run independently and manually, so the login_auth_test.py file has
been renamed to login_auth_test_must_run_manually.py in order to allow it
to continue to exist and be run, but not along with the rest of the suite.