[#600] Comment out test_files_query_case_sensitive BETWEEN tests
The BETWEEN clause behaves differently for mysql (and mariadb, I assume)
than postgres, and the test_files_query_case_sensitive test assumes that
we are using postgres. In order to avoid this, the BETWEEN sub-tests have
been commented out until such a time as we can make meaningful assertions
about the case-sensitive query with a BETWEEN clause based on the flavor
of database being used by the connected iRODS server.