[#3] tweak compatibility to iRODS 4.3.3
This change modifies the irods.message.IRODS_VERSION global variable
to bring the advertised level of server compatibility up-to-date with
release 4.3.3 of the iRODS Server.  The only obvious effect is in the
use of irods.test.helpers.make_session() to create session objects for
test purposes, as this call will raise an exception if the server is too
recent.

IRODS_VERSION is also reflected in the 'relVersion' and 'apiVersion' tags
of the StartupPack message when a session is initiated.