[_650] new escaped character processing
This fixes the approach to escaping characters from the set: [@&;=]
that iRODS historically has had problems accommodating in PAM passwords
in the past.  Presently, for iRODS 4.2 and 4.3 we only need to consider
";" and "=" as problematic characters, due to the conflict with the
use of those characters in the KVP-formatted context parameter when the
AUTH_PLUG_REQ_AN api is used.