[#657] catch Errno 57 on BSD/MacOS from socket.shutdown()
There is a race condition in BSD (of which MacOS is a decendent)
where socket.shutdown() only succeeds if the socket is still open.

If the socket is closed, the call fails with
OSError: [Errno 57] Socket is not connected

Linux and Windows do not error out if the socket is closed.