[#440] remove symbols that reference "user groups"
Originally, PRC models, objects, variables, etc., that had to do with groups of iRODS
users referred to them as "user groups."  The extra verbosity was intended to distinguish
them from the (no longer extant) resource groups.

In v1.1.7, when groupadmin capabilities were being upgraded, we also
introduced new symbol aliases that referred to a "UserGroups" as simply a "Group" and
deprecated the use of the old symbols. As a side benefit, some of the new symbols
were tailored to allow a "groupadmin" transparent access to their special
capabilites whenever the group_admin parameter was allowed to naturally default.)

The deprecated older symbols are hereby removed.