[#697] Github workflows for running PRC tests
Co-authored-by: d-w-moore <dmoore@renci.org>

Includes a docker compose configuration in which the PRC test suite
can be run in a way representative of typical operational use, and the
version of iRODS server and python interpreter that we install are also easily
reconfigurable.

Additionally, this affords us the opportunity to run the test suite on a client
node all its own, the iRODS server being reachable only via the network.