[#502] test harness and container-based tests
A new test harness is introduced in which we construct a new
container (using either Docker or podman) for each test program
we run. This allows full customization of the container
environment for the particular needs of each test.

Accordingly, also included with the Github workflows is a
full run of the PRC test suite with the iRODS server and
catalog DB server running in the same container as the client.

In the process of putting old tests through new rigors, faults
were found and corrected in some of those tests.